// src/common/enums/medical-topic.enum.ts
export enum MedicalTopicKey {
  MALARIA = 'MALARIA',
  TUBERCULOSIS = 'TUBERCULOSIS',
  DIABETES = 'DIABETES',
  HYPERTENSION = 'HYPERTENSION',
  SICKLE_CELL_DISEASE = 'SICKLE_CELL_DISEASE',
  CHRONIC_KIDNEY_DISEASE = 'CHRONIC_KIDNEY_DISEASE',
  DIALYSIS = 'DIALYSIS',
  VACCINATION = 'VACCINATION',
  NUTRITION = 'NUTRITION',
}

export enum MedicalTopicType {
  DISEASE = 'DISEASE',
  THEME = 'THEME',
}